API-217935 - Split main single file into separate class files.
